Remove some unnecessary ensure_can_withdraw (#248)
* Remove some unnecessary ensure_can_withdraw

* change AccountId to u128 for tests

+++ modules/loans/src/mock.rs
@@ -9,7 +9,7 @@ use sp_core::H256;
 use sp_runtime::{testing::Header, traits::IdentityLookup, Perbill};
 use support::{AuctionManager, RiskManager};
 
-pub type AccountId = u32;
+pub type AccountId = u128;
 pub type AuctionId = u64;
 pub type BlockNumber = u64;
 pub type DebitBalance = Balance;
